feat(password-service): refactor - moved classes
